Add stubs for non-Linux builds of statvfs and netlink.
This will allow the code to build when cross-compiled. The existing
functionality will be missing and we will degrade gracefully.
