+++ fs/statvfs_stub.go
@@ -1,0 +1,24 @@
+// Copyright 2015 Google Inc. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// +build !cgo,!linux
+
+package fs
+
+import "fmt"
+
+// Stub for unsupported builds.
+func getVfsStats(path string) (total uint64, free uint64, err error) {
+	return 0, 0, fmt.Errorf("vfsstats unsupported")
+}
