+++ utils/cpuload/netlink/defs_linux.go
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+// +build cgo,linux
+
 package netlink
 
 /*
